feat(api): implements the route to delete students from the database and add tests to it's service layer functions

+++ back-end/src/services/StudentsService.ts
@@ -48,6 +48,11 @@ class StudentsService {
   }
 
   public async deleteStudent(ra: string) {
+    if (!ra) throw new InvalidRequestError();
+
+    const studentExists = await this.studentsRepository.getStudent(ra);
+    if (!studentExists) throw new ResourceNotFound();
+
     return this.studentsRepository.deleteStudent(ra);
   }
 }
